requirements.txt
streamlit==1.32.0
pandas==2.2.1
Pillow==10.2.0
openpyxl==3.1.2
reportlab==4.1.0
python-dotenv==1.0.1
SQLAlchemy==2.0.27
mysql-connector-python==8.3.0 